+++ simvue/client.py
@@ -47,16 +47,13 @@ class Client(object):
                   'tags': tags,
                   'metadata': metadata}
 
-        try:
-            response = requests.get(f"{self._url}/api/runs", headers=self._headers, params=params)
-        except requests.exceptions.RequestException:
-            return None
+        response = requests.get(f"{self._url}/api/runs", headers=self._headers, params=params)
+        response.raise_for_status()
 
         if response.status_code == 200:
             return response.json()
 
         return None
-
 
     def get_runs(self, filters, system=False, tags=False, metadata=False, format='dict'):
         """
@@ -68,10 +65,8 @@ class Client(object):
                   'tags': tags,
                   'metadata': metadata}
 
-        try:
-            response = requests.get(f"{self._url}/api/runs", headers=self._headers, params=params)
-        except requests.exceptions.RequestException:
-            return None
+        response = requests.get(f"{self._url}/api/runs", headers=self._headers, params=params)
+        response.raise_for_status()
 
         if response.status_code == 200:
             if format == 'dict':
@@ -79,7 +74,7 @@ class Client(object):
             elif format == 'dataframe':
                 return to_dataframe(response.json())
             else:
-                return None 
+                raise Exception('invalid format specified')
 
         return None
 
@@ -91,15 +86,17 @@ class Client(object):
         if category:
             params['category'] = category        
 
-        try:
-            response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
-        except requests.exceptions.RequestException:
-            return None
+        response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
+
+        if response.status_code == 404:
+            if 'detail' in response.json():
+                if response.json()['detail'] == 'run does not exist':
+                    raise Exception('Run does not exist')
 
         if response.status_code == 200:
             return response.json()
 
-        return None
+        raise Exception(response.text)
 
     def get_artifact(self, run, name, allow_pickle=False):
         """
@@ -107,10 +104,14 @@ class Client(object):
         """
         params = {'run': run, 'name': name}
 
-        try:
-            response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
-        except requests.exceptions.RequestException:
-            return None
+        response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
+
+        if response.status_code == 404:
+            if 'detail' in response.json():
+                if response.json()['detail'] == 'run does not exist':
+                    raise Exception('Run does not exist')
+                elif response.json()['detail'] == 'artifact does not exist':
+                    raise Exception('Artifact does not exist')
 
         if response.status_code != 200:
             return None
@@ -118,10 +119,8 @@ class Client(object):
         url = response.json()[0]['url']
         mimetype = response.json()[0]['type']
 
-        try:
-            response = requests.get(url, timeout=DOWNLOAD_TIMEOUT)
-        except requests.exceptions.RequestException:
-            return None
+        response = requests.get(url, timeout=DOWNLOAD_TIMEOUT)
+        response.raise_for_status()
 
         content = Deserializer().deserialize(response.content, mimetype, allow_pickle)
         if content is not None:
@@ -135,10 +134,14 @@ class Client(object):
         """
         params = {'run': run, 'name': name}
 
-        try:
-            response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
-        except requests.exceptions.RequestException:
-            return None
+        response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
+
+        if response.status_code == 404:
+            if 'detail' in response.json():
+                if response.json()['detail'] == 'run does not exist':
+                    raise Exception('Run does not exist')
+                elif response.json()['detail'] == 'artifact does not exist':
+                    raise Exception('Artifact does not exist')
 
         if response.status_code == 200:
             if response.json():
@@ -146,6 +149,9 @@ class Client(object):
                 downloader({'url': url,
                             'filename': os.path.basename(name),
                             'path': path})
+
+        else:
+            raise Exception(response.text)
 
     def get_artifacts_as_files(self,
                                run,
@@ -161,10 +167,12 @@ class Client(object):
         if category:
             params['category'] = category
 
-        try:
-            response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
-        except requests.exceptions.RequestException:
-            return None
+        response = requests.get(f"{self._url}/api/artifacts", headers=self._headers, params=params)
+
+        if response.status_code == 404:
+            if 'detail' in response.json():
+                if response.json()['detail'] == 'run does not exist':
+                    raise Exception('Run does not exist')
 
         if not path:
             path = './'
@@ -199,3 +207,6 @@ class Client(object):
             with ProcessPoolExecutor(CONCURRENT_DOWNLOADS) as executor:
                 for item in downloads:
                     executor.submit(downloader, item)
+
+        else:
+            raise Exception(response.text)